requirements.txt
streamlit
tensorflow-cpu
numpy
pillow
opencv-python-headless